Comments done

+++ app/Http/Controllers/CommentController.php
@@ -14,12 +14,19 @@ class CommentController extends Controller
 
     public function store(Request $request)
     {
+        if (!auth()->user()) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
         $comment = new Comment();
         $comment->content = $request->content;
-        $comment->user_id = $request->user_id;
-        $comment->kebab_place_id = $request->kebab_place_id;
-
+        $comment->user_id = auth()->user()->id;
+        $comment->kebab_place_id = $request->kebabPlace;
         $comment->save();
+        return response()->json([
+            'message' => 'Comment created'
+        ]);
     }
 
     public function show(Comment $comment)
@@ -27,14 +34,42 @@ class CommentController extends Controller
         return json_encode($comment);
     }
 
-    public function update(Request $request, Comment $comment)
+    public function edit(Request $request)
     {
+        if (!auth()->user()) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+        $comment = Comment::find($request->comment);
+        if ($comment->user_id !== auth()->user()->id) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
         $comment->content ? $comment->content = $request->content : null;
         $comment->save();
+        return response()->json([
+            'message' => 'Comment edited'
+        ]);
     }
 
-    public function destroy(Comment $comment)
+    public function destroy(Request $request)
     {
+        if (!auth()->user()) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+        $comment = Comment::find($request->comment);
+        if ($comment->user_id !== auth()->user()->id) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
         $comment->delete();
+        return response()->json([
+            'message' => 'Comment deleted'
+        ]);
     }
 }
